Expose a sane implementation of LoadRows()
Previous implementation of LoadRows() didn't do what the caller
expected. Instead of refreshing the table with a new query result it
would reload the rows for whichever page the table was currently
displaying. If an app called LoadRows() when the page index was greater
than zero, and the query returned less than one page's worth of rows,
the result would be a table with blank rows. This commit makes the old
LoadRows() function private and replaces it with a "proper" LoadRows()
function that returns to the zero page and then re-executes the query.